tl_project/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-js-asset==3.1.2
django-mptt==0.16.0
environ==1.0
sqlparse==0.5.3
